MHC-7,8: hero details moved and refactored, tap from myavengers added
