+++ .idea/gradle.xml
@@ -15,6 +15,7 @@
             <option value="$PROJECT_DIR$/comics" />
             <option value="$PROJECT_DIR$/data" />
             <option value="$PROJECT_DIR$/domain" />
+            <option value="$PROJECT_DIR$/herodetails" />
             <option value="$PROJECT_DIR$/heroes" />
             <option value="$PROJECT_DIR$/myavengers" />
             <option value="$PROJECT_DIR$/settings" />
