+++ app/src/main/java/ru/coolhabit/marvelheroes/ioc/modules/FragmentsModule.kt
@@ -3,7 +3,7 @@ package ru.coolhabit.marvelheroes.ioc.modules
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 import ru.coolhabit.marvelheroes.heroes.presentation.base.HeroesFragment
-import ru.coolhabit.marvelheroes.heroes.presentation.detail.HeroDetailsFragment
+import ru.marvelheroes.herodetails.HeroDetailsFragment
 import ru.marvelheroes.comics.presentation.ComicsFragment
 import ru.marvelheroes.myavengers.presentation.MyAvengersFragment
 import ru.marvelheroes.settings.presentation.SettingsFragment
