+++ herodetails/src/main/java/ru/marvelheroes/herodetails/adapter/herodetail/HeroDetailSectionViewHolder.kt
@@ -1,34 +1,34 @@
-package ru.coolhabit.marvelheroes.heroes.presentation.detail.adapter.series
+package ru.marvelheroes.herodetails.adapter.herodetail
 
 import android.view.View
 import android.widget.TextView
 import androidx.annotation.LayoutRes
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import ru.coolhabit.marvelheroes.heroes.R
-import ru.coolhabit.marvelheroes.heroes.databinding.RvSectionSeriesBinding
-import ru.coolhabit.marvelheroes.heroes.presentation.detail.model.HeroDetailSection
+import ru.marvelheroes.herodetails.R
+import ru.marvelheroes.herodetails.databinding.RvSectionHeroCompilationBinding
+import ru.marvelheroes.herodetails.model.HeroDetailSection
 import ru.marvelheroes.presentation.adapter.HeroDetailBaseSectionViewHolder
 import ru.marvelheroes.presentation.adapter.IClickCommand
 import ru.marvelheroes.presentation.adapter.ItemDecoration
 
-class SeriesSectionViewHolder(view: View) :
-    HeroDetailBaseSectionViewHolder<HeroDetailSection.ComicsAndSeriesCompilation>(view) {
+class HeroDetailSectionViewHolder(view: View) :
+    HeroDetailBaseSectionViewHolder<HeroDetailSection.HeroDetailsCompilation>(view) {
 
     companion object {
         @LayoutRes
-        val ID: Int = R.layout.rv_section_series
+        val ID: Int = R.layout.rv_section_hero_compilation
     }
 
-    private val binding = RvSectionSeriesBinding.bind(view)
+    private val binding = RvSectionHeroCompilationBinding.bind(view)
 
     override val sectionName: TextView
         get() = binding.include.sectionName
     override val sectionList: RecyclerView
         get() = binding.include.viewContent
 
-    private val seriesAdapter: SeriesAdapter
-        get() = sectionList.adapter as SeriesAdapter
+    private val heroDetailAdapter: HeroDetailAdapter
+        get() = sectionList.adapter as HeroDetailAdapter
 
     override fun onCreated(callback: (IClickCommand<*>) -> Unit) {
         sectionList.apply {
@@ -37,19 +37,19 @@ class SeriesSectionViewHolder(view: View) :
             addItemDecoration(
                 ItemDecoration(
                     context,
-                    R.dimen.spacing_5,
-                    R.dimen.spacing_5,
-                    R.dimen.spacing_5,
-                    R.dimen.spacing_5
+                    R.dimen.size_0,
+                    R.dimen.size_0,
+                    R.dimen.size_0,
+                    R.dimen.size_0
                 )
             )
-            adapter = SeriesAdapter().apply {
+            adapter = HeroDetailAdapter().apply {
 
             }
         }
     }
 
-    override fun bind(item: HeroDetailSection.ComicsAndSeriesCompilation) {
-        seriesAdapter.submitList(item.bookList)
+    override fun bind(item: HeroDetailSection.HeroDetailsCompilation) {
+        heroDetailAdapter.submitList(item.heroDetails)
     }
 }
