+++ herodetails/src/main/java/ru/marvelheroes/herodetails/adapter/series/SeriesSectionViewHolder.kt
@@ -1,34 +1,34 @@
-package ru.coolhabit.marvelheroes.heroes.presentation.detail.adapter.herodetail
+package ru.marvelheroes.herodetails.adapter.series
 
 import android.view.View
 import android.widget.TextView
 import androidx.annotation.LayoutRes
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import ru.coolhabit.marvelheroes.heroes.R
-import ru.coolhabit.marvelheroes.heroes.databinding.RvSectionHeroCompilationBinding
-import ru.coolhabit.marvelheroes.heroes.presentation.detail.model.HeroDetailSection
+import ru.marvelheroes.herodetails.R
+import ru.marvelheroes.herodetails.databinding.RvSectionSeriesBinding
+import ru.marvelheroes.herodetails.model.HeroDetailSection
 import ru.marvelheroes.presentation.adapter.HeroDetailBaseSectionViewHolder
 import ru.marvelheroes.presentation.adapter.IClickCommand
 import ru.marvelheroes.presentation.adapter.ItemDecoration
 
-class HeroDetailSectionViewHolder(view: View) :
-    HeroDetailBaseSectionViewHolder<HeroDetailSection.HeroDetailsCompilation>(view) {
+class SeriesSectionViewHolder(view: View) :
+    HeroDetailBaseSectionViewHolder<HeroDetailSection.ComicsAndSeriesCompilation>(view) {
 
     companion object {
         @LayoutRes
-        val ID: Int = R.layout.rv_section_hero_compilation
+        val ID: Int = R.layout.rv_section_series
     }
 
-    private val binding = RvSectionHeroCompilationBinding.bind(view)
+    private val binding = RvSectionSeriesBinding.bind(view)
 
     override val sectionName: TextView
         get() = binding.include.sectionName
     override val sectionList: RecyclerView
         get() = binding.include.viewContent
 
-    private val heroDetailAdapter: HeroDetailAdapter
-        get() = sectionList.adapter as HeroDetailAdapter
+    private val seriesAdapter: SeriesAdapter
+        get() = sectionList.adapter as SeriesAdapter
 
     override fun onCreated(callback: (IClickCommand<*>) -> Unit) {
         sectionList.apply {
@@ -37,19 +37,19 @@ class HeroDetailSectionViewHolder(view: View) :
             addItemDecoration(
                 ItemDecoration(
                     context,
-                    R.dimen.size_0,
-                    R.dimen.size_0,
-                    R.dimen.size_0,
-                    R.dimen.size_0
+                    R.dimen.spacing_5,
+                    R.dimen.spacing_5,
+                    R.dimen.spacing_5,
+                    R.dimen.spacing_5
                 )
             )
-            adapter = HeroDetailAdapter().apply {
+            adapter = SeriesAdapter().apply {
 
             }
         }
     }
 
-    override fun bind(item: HeroDetailSection.HeroDetailsCompilation) {
-        heroDetailAdapter.submitList(item.heroDetails)
+    override fun bind(item: HeroDetailSection.ComicsAndSeriesCompilation) {
+        seriesAdapter.submitList(item.bookList)
     }
 }
