+++ herodetails/src/main/java/ru/marvelheroes/herodetails/model/HeroDetailsExt.kt
@@ -1,4 +1,4 @@
-package ru.coolhabit.marvelheroes.heroes.presentation.detail.model
+package ru.marvelheroes.herodetails.model
 
 import ru.marvelheroes.entities.dto.hero.HeroDetailData
