+++ herodetails/src/main/res/font/open_sans_family.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<font-family xmlns:android="http://schemas.android.com/apk/res/android">
+    <font
+        android:fontStyle="normal"
+        android:fontWeight="800"
+        android:font="@font/open_sans"/>
+</font-family>
