+++ herodetails/src/main/res/layout/fragment_hero_details.xml
