+++ herodetails/src/main/res/layout/rv_section_base.xml
