+++ herodetails/src/main/res/layout/view_hero_compilation.xml
