+++ herodetails/src/main/res/navigation/herodetails_graph.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/herodetails_graph"
+    app:startDestination="@id/fragment_hero_details">
+
+    <fragment
+        android:id="@+id/fragment_hero_details"
+        android:name="ru.marvelheroes.herodetails.HeroDetailsFragment"
+        tools:layout="@layout/fragment_hero_details">
+        <argument
+            android:name="heroId"
+            app:argType="string"/>
+        <deepLink
+            app:uri="@string/deeplink_hero_details">
+        </deepLink>
+    </fragment>
+</navigation>
