+++ heroes/build.gradle
@@ -80,4 +80,5 @@ dependencies {
 
     implementation project(':baseUI')
     implementation project(':domain')
+    implementation project(':herodetails')
 }
