+++ myavengers/build.gradle
@@ -78,4 +78,5 @@ dependencies {
 
     implementation project(':baseUI')
     implementation project(':domain')
+    implementation project(':herodetails')
 }
