+++ settings.gradle
@@ -21,3 +21,4 @@ include ':myavengers'
 include ':settings'
 include ':data'
 include ':domain'
+include ':herodetails'
